Split some EVM parts out to common
