+++ runtime/common/src/lib.rs
@@ -19,6 +19,7 @@
 mod tests;
 
 pub mod apis;
+pub mod evm;
 
 pub mod xcm_fees {
 	use cfg_primitives::{constants::currency_decimals, types::Balance};
